+++ R/config.R
@@ -28,6 +28,11 @@
 #' to certain duckdb functions.
 #' Currently unused.
 #'
+#' `DUCKPLYR_METHODS_OVERWRITE`: If `TRUE`, call `methods_overwrite()`
+#' when the package is loaded.
+#'
+#' See [fallback] for more options related to logging and uploading of fallback events.
+#'
 # Not available in the CRAN package:
 # `DUCKPLYR_META_GLOBAL`: Assume data frames in the global environment as "known".
 # `DUCKPLYR_META_SKIP`: Skip recording the operations, replay not available.