+++ R/duckplyr-package.R
@@ -4,30 +4,33 @@
 # @rawNamespace import(vctrs, except = data_frame)
 # an alternative for importing nearly everything from vctrs
 # https://github.com/tidyverse/dplyr/blob/16b472fb2afc50a87502c2b4ed803e2f5f82b9d6/R/dplyr.R#L7
+#
+# Can't use blanket cli import, imports must align with dplyr's imports
+# (except we can import all of dplyr)
+#
 #' @import rlang
 ## usethis namespace: start
 #' @importFrom collections dict
 #' @importFrom collections queue
+#' @importFrom dplyr group_by_prepare
+#' @importFrom dplyr group_indices
+#' @importFrom dplyr group_nest
+#' @importFrom dplyr group_split
+#' @importFrom dplyr group_trim
+#' @importFrom dplyr grouped_df
 #' @importFrom dplyr if_else
+#' @importFrom dplyr n_groups
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
-#' @importFrom purrr imap
-#' @importFrom purrr map
-#' @importFrom purrr map_chr
-#' @importFrom purrr map_lgl
-#' @importFrom purrr map2
-#' @importFrom purrr pmap
-#' @importFrom purrr reduce
-#' @importFrom purrr walk
 #' @importFrom tibble as_tibble
 #' @importFrom tibble deframe
 #' @importFrom tibble is_tibble
 #' @importFrom tibble new_tibble
 #' @importFrom tibble tibble
 #' @importFrom tidyselect everything
-#' @importFrom utils globalVariables
 #' @importFrom utils head
 #' @importFrom vctrs new_data_frame
+#' @importFrom vctrs new_list_of
 #' @importFrom vctrs new_rcrd
 #' @importFrom vctrs unspecified
 #' @importFrom vctrs vec_as_names
@@ -37,9 +40,11 @@
 #' @importFrom vctrs vec_cast_common
 #' @importFrom vctrs vec_cbind
 #' @importFrom vctrs vec_check_size
+#' @importFrom vctrs vec_data
 #' @importFrom vctrs vec_in
 #' @importFrom vctrs vec_match
 #' @importFrom vctrs vec_rbind
+#' @importFrom vctrs vec_recycle_common
 #' @importFrom vctrs vec_rep
 #' @importFrom vctrs vec_rep_each
 #' @importFrom vctrs vec_set_difference